pyproject.toml
[tool.poetry]
name = "llmlab"
version = "0.1.0"
description = ""
authors = ["vXofi <vladfaer10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.4"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.1.post1"
seaborn = "^0.13.2"
plotly = "^5.24.1"
torch = "^2.5.0"
torchvision = "^0.20.1"
tqdm = "^4.67.0"
llama-cpp-python = "^0.3.2"
huggingface-hub = "^0.26.2"
transformers = "^4.46.3"
sentence-transformers = "^3.3.1"
langchain = "^0.3.8"
duckduckgo-search = "^6.3.6"
langchain-community = "^0.3.8"
playwright = "^1.49.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
